Working backend with DB initialization through hibernate
Have to check functionality by implementing basic thymeleaf html files.

+++ Hotel/src/main/java/com/Manager/Hotel/repository/HotelRepository.java
@@ -5,7 +5,7 @@ import com.Manager.Hotel.entity.Hotel;
 
 import java.util.List;
 
-public interface HotelRepository extends JpaRepository<Hotel, Integer> {
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
 
     List<Hotel> findByName(String name);
