+++ Hotel/src/main/java/com/Manager/Hotel/service/CustomerService.java
@@ -22,11 +22,11 @@ public class CustomerService {
         return customerRepository.save(customer);
     }
 
-    public Customer getCustomer(int id) {
+    public Customer getCustomer(Long id) {
         return customerRepository.findById(id).orElse(null);
     }
 
-    public void deleteCustomer(int id) {
+    public void deleteCustomer(Long id) {
         customerRepository.deleteById(id);
     }
